editor/requirements.txt
Django==1.11.6
django-admin-sortable2==0.6.19
django-ckeditor==5.3.0
django-js-asset==0.1.1
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.2
Unidecode==0.4.21
